+++ socketcan_bridge/test/to_socketcan_test.cpp
@@ -63,8 +63,8 @@ TEST(TopicToSocketCANTest, checkCorrectData)
   // init the driver to test stateListener (not checked automatically).
   driver_->init("string_not_used", true);
 
-  // register for messages on received_messages.
-  ros::Publisher publisher_ = nh.advertise<can_msgs::Frame>("sent_messages", 10);
+  // register for messages on received_fd_messages.
+  ros::Publisher publisher_ = nh.advertise<can_msgs::FrameFd>("sent_fd_messages", 10);
 
   // create a frame collector.
   frameCollector frame_collector_;
@@ -75,7 +75,7 @@ TEST(TopicToSocketCANTest, checkCorrectData)
             std::bind(&frameCollector::frameCallback, &frame_collector_, std::placeholders::_1));
 
   // create a message
-  can_msgs::Frame msg;
+  can_msgs::FrameFd msg;
   msg.is_extended = true;
   msg.is_rtr = false;
   msg.is_error = false;
@@ -83,7 +83,7 @@ TEST(TopicToSocketCANTest, checkCorrectData)
   msg.dlc = 8;
   for (uint8_t i=0; i < msg.dlc; i++)
   {
-    msg.data[i] = i;
+    msg.data.push_back(i);
   }
 
   msg.header.frame_id = "0";  // "0" for no frame.
@@ -96,12 +96,71 @@ TEST(TopicToSocketCANTest, checkCorrectData)
   ros::WallDuration(1.0).sleep();
   ros::spinOnce();
 
-  can_msgs::Frame received;
+  can_msgs::FrameFd received;
   can::Frame f = frame_collector_.frames.back();
-  socketcan_bridge::convertSocketCANToMessage(f, received);
+  socketcan_bridge::convertSocketCANFDToMessage(f, received);
 
   EXPECT_EQ(received.id, msg.id);
   EXPECT_EQ(received.dlc, msg.dlc);
+  EXPECT_EQ(received.is_extended, msg.is_extended);
+  EXPECT_EQ(received.is_rtr, msg.is_rtr);
+  EXPECT_EQ(received.is_error, msg.is_error);
+  EXPECT_EQ(received.data, msg.data);
+}
+
+TEST(TopicToSocketCANTest, checkCorrectFdData)
+{
+  ros::NodeHandle nh(""), nh_param("~");
+
+  // create the dummy interface
+  can::DummyInterfaceSharedPtr driver_ = std::make_shared<can::DummyInterface>(true);
+
+  // start the to topic bridge.
+  socketcan_bridge::TopicToSocketCAN to_socketcan_bridge(&nh, &nh_param, driver_);
+  to_socketcan_bridge.setup();
+
+  // init the driver to test stateListener (not checked automatically).
+  driver_->init("string_not_used", true);
+
+  // register for messages on received_fd_messages.
+  ros::Publisher publisher_ = nh.advertise<can_msgs::FrameFd>("sent_fd_messages", 10);
+
+  // create a frame collector.
+  frameCollector frame_collector_;
+
+  //  driver->createMsgListener(&frameCallback);
+  can::FrameListenerConstSharedPtr frame_listener_ = driver_->createMsgListener(
+
+            std::bind(&frameCollector::frameCallback, &frame_collector_, std::placeholders::_1));
+
+  // create a message
+  can_msgs::FrameFd msg;
+  msg.is_extended = true;
+  msg.is_rtr = false;
+  msg.is_error = false;
+  msg.id = 0x123;
+  msg.dlc = 64 | can_msgs::FrameFd::DLC_FD_BRS_FLAG;
+  for (uint8_t i=0; i < 64; i++)
+  {
+    msg.data.push_back(i);
+  }
+
+  msg.header.frame_id = "0";  // "0" for no frame.
+  msg.header.stamp = ros::Time::now();
+
+  // send the can_frame::Frame message to the sent_messages topic.
+  publisher_.publish(msg);
+
+  // give some time for the interface some time to process the message
+  ros::WallDuration(1.0).sleep();
+  ros::spinOnce();
+
+  can_msgs::FrameFd received;
+  can::Frame f = frame_collector_.frames.back();
+  socketcan_bridge::convertSocketCANFDToMessage(f, received);
+
+  EXPECT_EQ(received.id, msg.id);
+  EXPECT_EQ(received.dlc, 64);
   EXPECT_EQ(received.is_extended, msg.is_extended);
   EXPECT_EQ(received.is_rtr, msg.is_rtr);
   EXPECT_EQ(received.is_error, msg.is_error);
@@ -125,8 +184,8 @@ TEST(TopicToSocketCANTest, checkInvalidFrameHandling)
   socketcan_bridge::TopicToSocketCAN to_socketcan_bridge(&nh, &nh_param, driver_);
   to_socketcan_bridge.setup();
 
-  // register for messages on received_messages.
-  ros::Publisher publisher_ = nh.advertise<can_msgs::Frame>("sent_messages", 10);
+  // register for messages on received_fd_messages.
+  ros::Publisher publisher_ = nh.advertise<can_msgs::FrameFd>("sent_fd_messages", 10);
 
   // create a frame collector.
   frameCollector frame_collector_;
@@ -136,7 +195,7 @@ TEST(TopicToSocketCANTest, checkInvalidFrameHandling)
           std::bind(&frameCollector::frameCallback, &frame_collector_, std::placeholders::_1));
 
   // create a message
-  can_msgs::Frame msg;
+  can_msgs::FrameFd msg;
   msg.is_extended = false;
   msg.id = (1<<11)+1;  // this is an illegal CAN packet... should not be sent.
   msg.header.frame_id = "0";  // "0" for no frame.
@@ -162,8 +221,8 @@ TEST(TopicToSocketCANTest, checkInvalidFrameHandling)
   frame_collector_.frames.clear();
 
 
-  // finally, check if frames with a dlc > 8 are discarded.
-  msg.dlc = 10;
+  // finally, check if frames with a dlc > 64 are discarded.
+  msg.dlc = 65;
   publisher_.publish(msg);
   ros::WallDuration(1.0).sleep();
   ros::spinOnce();