+++ socketcan_bridge/test/to_topic_test.cpp
@@ -27,6 +27,7 @@
 #include <socketcan_bridge/socketcan_to_topic.h>
 
 #include <can_msgs/Frame.h>
+#include <can_msgs/FrameFd.h>
 #include <socketcan_interface/socketcan.h>
 #include <socketcan_interface/dummy.h>
 #include <socketcan_bridge/topic_to_socketcan.h>
@@ -42,6 +43,7 @@ class msgCollector
 {
   public:
     std::list<can_msgs::Frame> messages;
+    std::list<can_msgs::FrameFd> messages_fd;
 
     msgCollector() {}
 
@@ -49,12 +51,24 @@ class msgCollector
     {
       messages.push_back(f);
     }
+
+    void msgFdCallback(const can_msgs::FrameFd& f)
+    {
+      messages_fd.push_back(f);
+    }
 };
 
 std::string convertMessageToString(const can_msgs::Frame &msg, bool lc = true)
 {
   can::Frame f;
   socketcan_bridge::convertMessageToSocketCAN(msg, f);
+  return can::tostring(f, lc);
+}
+
+std::string convertFdMessageToString(const can_msgs::FrameFd &msg, bool lc = true)
+{
+  can::Frame f;
+  socketcan_bridge::convertMessageToSocketCANFD(msg, f);
   return can::tostring(f, lc);
 }
 
@@ -80,6 +94,7 @@ TEST(SocketCANToTopicTest, checkCorrectData)
 
   // create a can frame
   can::Frame f;
+  f.is_fd = false;
   f.is_extended = true;
   f.is_rtr = false;
   f.is_error = false;
@@ -99,17 +114,89 @@ TEST(SocketCANToTopicTest, checkCorrectData)
 
   ASSERT_EQ(1, message_collector_.messages.size());
 
-  // compare the received can_msgs::Frame message to the sent can::Frame.
+  // compare the received can_msgs::FrameFd message to the sent can::Frame.
   can::Frame received;
   can_msgs::Frame msg = message_collector_.messages.back();
   socketcan_bridge::convertMessageToSocketCAN(msg, received);
+
+  can::Frame received_fd;
+  can_msgs::FrameFd msg_fd = message_collector_.messages_fd.back();
+  socketcan_bridge::convertMessageToSocketCANFD(msg_fd, received_fd);
 
   EXPECT_EQ(received.id, f.id);
   EXPECT_EQ(received.dlc, f.dlc);
   EXPECT_EQ(received.is_extended, f.is_extended);
   EXPECT_EQ(received.is_rtr, f.is_rtr);
   EXPECT_EQ(received.is_error, f.is_error);
-  EXPECT_EQ(received.data, f.data);
+  for (int i = 0; i < f.dlc; ++i)
+  {
+    EXPECT_EQ(received.data[i], f.data[i]);
+
+    // just for debug
+    EXPECT_EQ(msg.data[i], f.data[i]);
+  }
+}
+
+TEST(SocketCANToTopicTest, checkCorrectFdData)
+{
+  ros::NodeHandle nh(""), nh_param("~");
+
+  // create the dummy interface
+  can::DummyInterfaceSharedPtr driver_ = std::make_shared<can::DummyInterface>(true);
+
+  // start the to topic bridge.
+  socketcan_bridge::SocketCANToTopic to_topic_bridge(&nh, &nh_param, driver_);
+  to_topic_bridge.setup();  // initiate the message callbacks
+
+  // init the driver to test stateListener (not checked automatically).
+  driver_->init("string_not_used", true);
+
+  // create a frame collector.
+  msgCollector message_collector_;
+
+  // register for messages on received_fd_messages.
+  ros::Subscriber subscriber_ = nh.subscribe("received_fd_messages", 1,
+                                             &msgCollector::msgFdCallback,
+                                             &message_collector_);
+
+  // create a can frame
+  can::Frame f;
+  f.is_fd = true;
+  f.is_extended = true;
+  f.is_rtr = false;
+  f.is_error = false;
+  f.id = 0x123;
+  f.dlc = 64;
+  f.is_fd = true;
+  for (uint8_t i=0; i < f.dlc; i++)
+  {
+    f.data[i] = i;
+  }
+
+  // send the can frame to the driver
+  driver_->send(f);
+
+  // give some time for the interface some time to process the message
+  ros::WallDuration(1.0).sleep();
+  ros::spinOnce();
+
+  ASSERT_EQ(1, message_collector_.messages_fd.size());
+
+  // compare the received can_msgs::FrameFd message to the sent can::Frame.
+  can::Frame received;
+  can_msgs::FrameFd msg = message_collector_.messages_fd.back();
+  socketcan_bridge::convertMessageToSocketCANFD(msg, received);
+
+  EXPECT_EQ(received.id, f.id);
+  EXPECT_EQ(received.dlc, f.dlc);
+  EXPECT_EQ(received.is_extended, f.is_extended);
+  EXPECT_EQ(received.is_rtr, f.is_rtr);
+  EXPECT_EQ(received.is_error, f.is_error);
+  EXPECT_EQ(received.is_fd, f.is_fd);
+  for (int i = 0; i < f.dlc; ++i)
+  {
+    EXPECT_EQ(received.data[i], f.data[i]);
+  }
 }
 
 TEST(SocketCANToTopicTest, checkInvalidFrameHandling)
@@ -118,7 +205,7 @@ TEST(SocketCANToTopicTest, checkInvalidFrameHandling)
   //   that should not be sent.
   // - verifies that sending one larger than 11 bits actually works.
 
-  // sending a message with a dlc > 8 is not possible as the DummyInterface
+  // sending a message with a dlc > 64 is not possible as the DummyInterface
   // causes a crash then.
 
   ros::NodeHandle nh(""), nh_param("~");
@@ -133,29 +220,33 @@ TEST(SocketCANToTopicTest, checkInvalidFrameHandling)
   // create a frame collector.
   msgCollector message_collector_;
 
-  // register for messages on received_messages.
-  ros::Subscriber subscriber_ = nh.subscribe("received_messages", 1, &msgCollector::msgCallback, &message_collector_);
+  // register for messages on received_fd_messages.
+  ros::Subscriber subscriber_ = nh.subscribe("received_fd_messages", 1,
+                                             &msgCollector::msgFdCallback,
+                                             &message_collector_);
 
   // create a message
   can::Frame f;
+  f.is_fd = true;
   f.is_extended = false;
   f.id = (1<<11)+1;  // this is an illegal CAN packet... should not be sent.
 
   // send the can::Frame over the driver.
-  // driver_->send(f);
+  driver_->send(f);
 
   // give some time for the interface some time to process the message
   ros::WallDuration(1.0).sleep();
   ros::spinOnce();
-  EXPECT_EQ(message_collector_.messages.size(), 0);
+  EXPECT_EQ(message_collector_.messages_fd.size(), 0);
 
+  f.is_fd = true;
   f.is_extended = true;
   f.id = (1<<11)+1;  // now it should be alright.
 
   driver_->send(f);
   ros::WallDuration(1.0).sleep();
   ros::spinOnce();
-  EXPECT_EQ(message_collector_.messages.size(), 1);
+  EXPECT_EQ(message_collector_.messages_fd.size(), 1);
 }
 
 TEST(SocketCANToTopicTest, checkCorrectCanIdFilter)
@@ -179,11 +270,12 @@ TEST(SocketCANToTopicTest, checkCorrectCanIdFilter)
   // create a frame collector.
   msgCollector message_collector_;
 
-  // register for messages on received_messages.
+  // register for messages on received_fd_messages.
   ros::Subscriber subscriber_ = nh.subscribe("received_messages", 1, &msgCollector::msgCallback, &message_collector_);
 
   // create a can frame
   can::Frame f;
+  f.is_fd = false;
   f.is_extended = true;
   f.is_rtr = false;
   f.is_error = false;
@@ -203,7 +295,7 @@ TEST(SocketCANToTopicTest, checkCorrectCanIdFilter)
 
   ASSERT_EQ(1, message_collector_.messages.size());
 
-  // compare the received can_msgs::Frame message to the sent can::Frame.
+  // compare the received can_msgs::FrameFd message to the sent can::Frame.
   can::Frame received;
   can_msgs::Frame msg = message_collector_.messages.back();
   socketcan_bridge::convertMessageToSocketCAN(msg, received);
@@ -213,7 +305,10 @@ TEST(SocketCANToTopicTest, checkCorrectCanIdFilter)
   EXPECT_EQ(received.is_extended, f.is_extended);
   EXPECT_EQ(received.is_rtr, f.is_rtr);
   EXPECT_EQ(received.is_error, f.is_error);
-  EXPECT_EQ(received.data, f.data);
+  for (int i = 0; i < f.dlc; ++i)
+  {
+    EXPECT_EQ(received.data[i], f.data[i]);
+  }
 }
 
 TEST(SocketCANToTopicTest, checkInvalidCanIdFilter)
@@ -237,11 +332,14 @@ TEST(SocketCANToTopicTest, checkInvalidCanIdFilter)
   // create a frame collector.
   msgCollector message_collector_;
 
-  // register for messages on received_messages.
-  ros::Subscriber subscriber_ = nh.subscribe("received_messages", 1, &msgCollector::msgCallback, &message_collector_);
+  // register for messages on received_fd_messages.
+  ros::Subscriber subscriber_ = nh.subscribe("received_fd_messages", 1,
+                                             &msgCollector::msgFdCallback,
+                                             &message_collector_);
 
   // create a can frame
   can::Frame f;
+  f.is_fd = true;
   f.is_extended = true;
   f.is_rtr = false;
   f.is_error = false;
@@ -283,7 +381,7 @@ TEST(SocketCANToTopicTest, checkMaskFilter)
   // create a frame collector.
   msgCollector message_collector_;
 
-  // register for messages on received_messages.
+  // register for messages on received_fd_messages.
   ros::Subscriber subscriber_ = nh.subscribe("received_messages", 10, &msgCollector::msgCallback, &message_collector_);
 
   const std::string pass1("300#1234"), nopass1("302#9999"), pass2("301#5678");